add modal dialog fixes #633 and mark for Lifter 1 refs #68

+++ public/contact_statusgruppen.php
@@ -1,4 +1,5 @@
 <?
+# Lifter001: TODO
 # Lifter002: TODO
 /*
 admin_statusgruppe.php - Statusgruppen-Verwaltung von Stud.IP.
@@ -95,7 +96,7 @@ function PrintAktualStatusgruppen ($range_id, $view, $edit_id="")
 		echo "\n\t\t<td width=\"5%\">";
 		printf ("            	  <input type=\"IMAGE\" name=\"%s\" src=\"".$GLOBALS['ASSETS_URL']."images/move.gif\" border=\"0\" %s>&nbsp; </td>", $statusgruppe_id, tooltip(_("Markierte Personen dieser Gruppe zuordnen")));
 		printf ("	          <td width=\"85%%\" class=\"%s\">&nbsp; %s </td><td class=\"%s\" width=\"5%%\"><a href=\"$PHP_SELF?cmd=edit_statusgruppe&edit_id=%s&range_id=%s&view=%s\"><img src=\"".$GLOBALS['ASSETS_URL']."images/einst.gif\" border=\"0\" %s></a></td>",$edit_id == $statusgruppe_id?"topicwrite":"topic", htmlReady($db->f("name")), $edit_id == $statusgruppe_id?"topicwrite":"topic", $statusgruppe_id, $range_id, $view, tooltip(_("Gruppenname oder -größe anpassen")));
-		printf ( "	          <td width=\"5%%\"><a href=\"$PHP_SELF?cmd=remove_statusgruppe&statusgruppe_id=%s&range_id=%s&view=%s\"><img src=\"".$GLOBALS['ASSETS_URL']."images/trash_att.gif\" width=\"11\" height=\"17\" border=\"0\" %s></a></td>",$statusgruppe_id, $range_id, $view, tooltip(_("Gruppe mit Personenzuordnung entfernen")));
+		printf ( "	          <td width=\"5%%\"><a href=\"$PHP_SELF?cmd=verify_remove_statusgruppe&statusgruppe_id=%s&range_id=%s&view=%s&name=%s\"><img src=\"".$GLOBALS['ASSETS_URL']."images/trash_att.gif\" width=\"11\" height=\"17\" border=\"0\" %s></a></td>",$statusgruppe_id, $range_id, $view, $db->f('name'), tooltip(_("Gruppe mit Personenzuordnung entfernen")));
 		echo 	"\n\t</tr>";
 
 		$db2->query ("SELECT statusgruppe_user.user_id, " . $_fullname_sql['full'] . " AS fullname , username FROM statusgruppe_user LEFT JOIN auth_user_md5 USING(user_id) LEFT JOIN user_info USING (user_id) WHERE statusgruppe_id = '$statusgruppe_id'");
@@ -226,10 +227,15 @@ function PrintAktualContacts ($range_id)
 
 	// Entfernen von Statusgruppen
 
-	if ($cmd=="remove_statusgruppe") {
-		DeleteStatusgruppe ($statusgruppe_id);
+	if ($cmd=="verify_remove_statusgruppe") {
+		$msg = sprintf(_('Möchten Sie wirklich die Kategorie **%s** löschen?'), $name); 
+		echo createQuestion($msg, array('cmd' => 'remove_statusgruppe', "statusgruppe_id" => $statusgruppe_id, 'range_id' => $range_id)); 
 	}
 
+    if ($cmd=="remove_statusgruppe") {
+		
+		DeleteStatusgruppe ($statusgruppe_id);
+	}
 	// Aendern der Position
 
 	if ($cmd=="swap") {
